refactor - Move PhoneNumberSerializer to common model
The PhoneNumber model is in the common app, so moved the serializer
as well (and also related tests) to common app
