+++ papapay/common/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .phone_number_serializer import PhoneNumberSerializer
+
+__all__ = [PhoneNumberSerializer]
