+++ papapay/user/serializers/__init__.py
@@ -1,10 +1,9 @@
 from .login_serializer import LoginSerializer
 from .password_update_serializer import PasswordUpdateSerializer
-from .phone_number_serializer import PhoneNumberSerializer
 from .signup_serializer import SignupSerializer
 from .user_profile_serializer import (PhoneNumberChoice,
                                       PhoneNumbersChoiceField,
                                       UserProfileSerializer)
 
-__all__ = [LoginSerializer, PasswordUpdateSerializer, PhoneNumberChoice, PhoneNumbersChoiceField,
-           PhoneNumberSerializer, SignupSerializer, UserProfileSerializer]
+__all__ = [LoginSerializer, PasswordUpdateSerializer, PhoneNumberChoice,
+           PhoneNumbersChoiceField, SignupSerializer, UserProfileSerializer]
